requirements.txt
aiohttp==3.9.3 ; python_version >= "3.10" and python_version < "3.12"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "3.12"
asn1crypto==1.5.1 ; python_version >= "3.10" and python_version < "3.12"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.11"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "3.12"
blinker==1.7.0 ; python_version >= "3.10" and python_version < "3.12"
cachetools==5.3.2 ; python_version >= "3.10" and python_version < "3.12"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "3.12"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "3.12" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.12"
cloud-sql-python-connector==1.6.0 ; python_version >= "3.10" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.12" and platform_system == "Windows"
cryptography==42.0.2 ; python_version >= "3.10" and python_version < "3.12"
flask==3.0.2 ; python_version >= "3.10" and python_version < "3.12"
frozenlist==1.4.1 ; python_version >= "3.10" and python_version < "3.12"
google-auth==2.27.0 ; python_version >= "3.10" and python_version < "3.12"
greenlet==3.0.3 ; python_version >= "3.10" and python_version < "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "3.12"
idna==3.6 ; python_version >= "3.10" and python_version < "3.12"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "3.12"
jinja2==3.1.3 ; python_version >= "3.10" and python_version < "3.12"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "3.12"
multidict==6.0.5 ; python_version >= "3.10" and python_version < "3.12"
numpy==1.26.4 ; python_version >= "3.10" and python_version <= "3.11"
packaging==23.2 ; python_version >= "3.10" and python_version < "3.12"
pandas==2.2.0 ; python_version >= "3.10" and python_version < "3.12"
pg8000==1.30.4 ; python_version >= "3.10" and python_version < "3.12"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "3.12"
pyasn1==0.5.1 ; python_version >= "3.10" and python_version < "3.12"
pycparser==2.21 ; python_version >= "3.10" and python_version < "3.12" and platform_python_implementation != "PyPy"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.12"
pytz==2024.1 ; python_version >= "3.10" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.12"
rsa==4.9 ; python_version >= "3.10" and python_version < "3.12"
scramp==1.4.4 ; python_version >= "3.10" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.12"
sqlalchemy==2.0.25 ; python_version >= "3.10" and python_version < "3.12"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "3.12"
tzdata==2023.4 ; python_version >= "3.10" and python_version < "3.12"
urllib3==2.2.0 ; python_version >= "3.10" and python_version < "3.12"
werkzeug==3.0.1 ; python_version >= "3.10" and python_version < "3.12"
yarl==1.9.4 ; python_version >= "3.10" and python_version < "3.12"
